+++ frontend/test/utils/numericHelpers.test.ts
@@ -3,11 +3,11 @@
  * Web client of the Sensor Management System software developed within
  * the Helmholtz DataHub Initiative by GFZ and UFZ.
  *
- * Copyright (C) 2020 - 2024
+ * Copyright (C) 2024
  * - Nils Brinckmann (GFZ, nils.brinckmann@gfz-potsdam.de)
  * - Marc Hanisch (GFZ, marc.hanisch@gfz-potsdam.de)
- * - Helmholtz Centre Potsdam - GFZ German Research Centre for
- *   Geosciences (GFZ, https://www.gfz-potsdam.de)
+ * - Helmholtz Centre Potsdam - GFZ German Research Centre for Geosciences
+ *   (GFZ, https://www.gfz-potsdam.de)
  *
  * Parts of this program were developed within the context of the
  * following publicly funded projects or measures:
@@ -28,9 +28,23 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
  * implied. See the Licence for the specific language governing
  * permissions and limitations under the Licence.
- */
-export const parseFloatOrDefault = (value: any, defaultValue: number): number => isNaN(parseFloat(value)) ? defaultValue : parseFloat(value)
-export const parseFloatOrNull = (value: any): number | null => isNaN(parseFloat(value)) ? null : parseFloat(value)
-export const parseIntOrDefault = (value: any, defaultValue: number): number => isNaN(parseInt(value)) ? defaultValue : parseInt(value)
-export const parseIntOrNull = (value: any): number | null => isNaN(parseInt(value)) ? null : parseInt(value)
-export const round = (value: number, ndigits: number = 0) => Math.round(value * 10 ** ndigits) / 10 ** ndigits
+*/
+
+import { round } from '@/utils/numericsHelper'
+
+describe('round', () => {
+  it('should use a default argument of 0', () => {
+    expect(round(1.4)).toEqual(1)
+  })
+  it('should be possible to set the ndigits', () => {
+    expect(round(1.4, 1)).toEqual(1.4)
+  })
+  it('should only take the part the point into account', () => {
+    expect(round(1234.35, 1)).toEqual(1234.4)
+  })
+  it('should help with odd arithmetic issues', () => {
+    const value = 2.7 - 1.0
+    expect(value).not.toEqual(1.7)
+    expect(round(value, 6)).toEqual(1.7)
+  })
+})